src/pyproject.toml
[project]
name = "tda-env"
requires-python = "~=3.11"
version = "0.0.0"

dependencies = [
    "bandit[toml]~=1.7",
    # "cyclopts~=2.9",
    # "fasttext-wheel~=0.9",
    "jupyter~=1.0",
    # # "ipykernel~=6.0",
    "matplotlib~=3.9",
    "mypy~=1.5",
    # "pandas~=2.0",
    "pandas",
    "pandas-stubs~=2.0",
    # "pytest~=8.0",
    # "pytest-cov~=5.0",
    # "python-dotenv~=1.0",
    "ruff~=0.7.0",
    # "seaborn~=0.12",
    # "POT~=0.9",
    # "scipy>=1.0",
    # "numpy~=1.17",
    # "numpy~=2.0",
    "numpy",
    # "cython>=0.23",
    # "opencv-python~=4.11",
    "scikit-learn",
    "umap-learn~=0.5",
    "kmapper>=2.0",
    "pyarrow~=20.0",
    "giotto-tda~=0.6.2", 
    "tda-mapper~=0.12",
    "ott-jax",
]

[build-system]
requires = ["setuptools~=78.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
exclude = [
    "build",
    "venv",
    "notebooks",
    "data",
    "venv_test"
]

[project.scripts]
optimal-transport = "optimal_transport.__main__:app"

[tool.ruff]
exclude = [
    "build",
    "venv",
    "notebooks",
    "data",
]

[tool.mypy]
exclude = [
    "build",
    "venv",
    "notebooks",
    "data",
]
ignore_missing_imports = true

[tool.bandit]
exclude_dirs = [
    "build",
    "venv",
    "notebooks",
    "data",
]

[tool.bandit.assert_used]
skips = ["**/test_*.py"]

[tool.pytest.ini_options]
log_cli = true
filterwarnings = [
    'ignore:FUNCTIONS is deprecated and will be removed in future versions:DeprecationWarning',
]
